requirements.txt
pytorch-lightning
beautifulsoup4
Django==3.1.3
matplotlib
opencv-python
Pillow
scikit-learn
scipy
tqdm
boto3
dominate>=2.3.1
dill
scikit-image
legacy
autopep8
beautifulsoup4
black
h5py
pandas
protobuf
pylint
pymongo
tensorboard
torch-encoding
torchmetrics
yapf